ci: forge fmt + sync addresses.ts + slither via pip
Three small CI fixes:

- Smart Contracts (forge fmt --check): ran `forge fmt` across src/,
  script/, test/ — 100+ files reformatted, no logic changes.
- Address Registry: ran `pnpm sync:addresses`, regenerated
  deployments/{sepolia,base-sepolia}.json and
  packages/abis/src/addresses.ts + apps/web/src/configs/addresses.ts
  from live on-chain state (27 Sepolia + 31 Base Sepolia contracts).
- Contract Security (Slither): replaced crytic/slither-action@v0.4.0
  with `pip3 install slither-analyzer` + a direct `slither .` call.
  The action's Docker image runs `npm install` at the repo root and
  crashes on pnpm's `workspace:*` protocol; the action's ignore-compile
  option then errors with `KeyError: 'output'` because it still
  expects crytic-compile artifacts. Running slither directly against
  the foundry `out/` tree sidesteps both.

+++ apps/contracts/src/Universe.sol
@@ -9,21 +9,18 @@ import {ReentrancyGuard} from "solady/src/utils/ReentrancyGuard.sol";
 import {Pausable} from "@openzeppelin/utils/Pausable.sol";
 
 contract Universe is IUniverse, ReentrancyGuard, Pausable {
-
     struct VideoNode {
-        bytes32 contentHash;   // SHA-256 hash of media file
-        uint id;
-        bytes32 plotHash;      // SHA-256 hash of plot text
-        uint previous;
-        uint[] next;
+        bytes32 contentHash; // SHA-256 hash of media file
+        uint256 id;
+        bytes32 plotHash; // SHA-256 hash of plot text
+        uint256 previous;
+        uint256[] next;
         /// @dev Only meaningful for currentCanonId. Use getCanonChain() to determine full canon membership.
         bool canon;
         address creator;
     }
 
-    constructor(
-        IUniverseManager.UniverseConfig memory config
-    ) {
+    constructor(IUniverseManager.UniverseConfig memory config) {
         require(config.universeAdmin != address(0), "Zero admin address");
         require(config.universeManager != address(0), "Zero manager address");
         nodeCreationOption = config.nodeCreationOption;
@@ -38,8 +35,8 @@ contract Universe is IUniverse, ReentrancyGuard, Pausable {
     string public universeImageUrl;
     string public universeName;
     string public universeDescription;
-    mapping(uint => VideoNode) public nodes;
-    uint public latestNodeId;
+    mapping(uint256 => VideoNode) public nodes;
+    uint256 public latestNodeId;
     mapping(address user => bool) isWhitelisted;
     mapping(address user => bool) public vaultWhitelisted;
 
@@ -49,34 +46,34 @@ contract Universe is IUniverse, ReentrancyGuard, Pausable {
     address public associatedToken;
     IUniverseManager public immutable universeManager;
     address public universeAdmin;
-    uint public currentCanonId;  // current canon tip node ID (not "is part of canon chain" — use getCanonChain())
+    uint256 public currentCanonId; // current canon tip node ID (not "is part of canon chain" — use getCanonChain())
 
     /// @notice Maximum children per node (prevents unbounded array growth)
-    uint public constant MAX_CHILDREN_PER_NODE = 100;
+    uint256 public constant MAX_CHILDREN_PER_NODE = 100;
 
     modifier onlyAdmin() {
-      _checkAdmin();
-      _;
+        _checkAdmin();
+        _;
     }
 
     function _checkAdmin() internal view {
-      if (universeAdmin != msg.sender) {
-        revert CallerNotAdmin(msg.sender);
-      }
+        if (universeAdmin != msg.sender) {
+            revert CallerNotAdmin(msg.sender);
+        }
     }
 
     modifier onlyManager() {
-      _checkManager();
-      _;
+        _checkManager();
+        _;
     }
 
     function _checkManager() internal view {
-      if (address(universeManager) != msg.sender) {
-        revert CallerNotManager();
-      }
+        if (address(universeManager) != msg.sender) {
+            revert CallerNotManager();
+        }
     }
 
-    function nodeIdToHex(uint id) public view returns (bytes32){
+    function nodeIdToHex(uint256 id) public view returns (bytes32) {
         if (id == 0 || id > latestNodeId) {
             revert NodeDoesNotExist();
         }
@@ -94,7 +91,7 @@ contract Universe is IUniverse, ReentrancyGuard, Pausable {
     /// @param status Whitelist status to set for all addresses
     function batchSetWhitelisted(address[] calldata users, bool status) external onlyAdmin {
         require(users.length <= 200, "Batch too large");
-        for (uint i = 0; i < users.length; i++) {
+        for (uint256 i = 0; i < users.length; i++) {
             isWhitelisted[users[i]] = status;
             emit WhitelistedUpdated(users[i], status);
         }
@@ -112,7 +109,7 @@ contract Universe is IUniverse, ReentrancyGuard, Pausable {
     /// @notice Batch vault whitelist multiple addresses in a single transaction.
     function batchSetVaultWhitelisted(address[] calldata users, bool status) external onlyAdmin {
         require(users.length <= 200, "Batch too large");
-        for (uint i = 0; i < users.length; i++) {
+        for (uint256 i = 0; i < users.length; i++) {
             vaultWhitelisted[users[i]] = status;
             emit VaultWhitelistUpdated(users[i], status);
         }
@@ -131,23 +128,21 @@ contract Universe is IUniverse, ReentrancyGuard, Pausable {
     function createNode(
         bytes32 _contentHash,
         bytes32 _plotHash,
-        uint _previous,
+        uint256 _previous,
         string calldata _link,
         string calldata _plot
-    ) public nonReentrant whenNotPaused returns (uint) {
+    ) public nonReentrant whenNotPaused returns (uint256) {
         if (nodeCreationOption == NodeCreationOptions.WHITELISTED) {
             require(isWhitelisted[msg.sender], "Not whitelisted");
         }
         require(_previous == 0 || nodes[_previous].id != 0, "Previous node does not exist");
         // Token-gated creation: if visibility is HOLDERS, require token balance
-        if (nodeVisibilityOption == NodeVisibilityOptions.HOLDERS && associatedToken != address(0)) {
-            require(
-                IERC20(associatedToken).balanceOf(msg.sender) > 0,
-                "Must hold universe token"
-            );
+        if (nodeVisibilityOption == NodeVisibilityOptions.HOLDERS && associatedToken != address(0))
+        {
+            require(IERC20(associatedToken).balanceOf(msg.sender) > 0, "Must hold universe token");
         }
         latestNodeId++;
-        uint newId = latestNodeId;
+        uint256 newId = latestNodeId;
 
         nodes[newId].id = newId;
         nodes[newId].contentHash = _contentHash;
@@ -168,40 +163,30 @@ contract Universe is IUniverse, ReentrancyGuard, Pausable {
         return newId;
     }
 
-    function getNode(
-        uint id
-    )
+    function getNode(uint256 id)
         public
         view
-        returns (
-            uint,
-            bytes32,
-            bytes32,
-            uint,
-            uint[] memory,
-            bool,
-            address
-        )
+        returns (uint256, bytes32, bytes32, uint256, uint256[] memory, bool, address)
     {
         VideoNode storage n = nodes[id];
         return (n.id, n.contentHash, n.plotHash, n.previous, n.next, n.canon, n.creator);
     }
 
     /// @notice Maximum timeline traversal depth to prevent gas exhaustion DoS
-    uint public constant MAX_TIMELINE_DEPTH = 1000;
+    uint256 public constant MAX_TIMELINE_DEPTH = 1000;
 
-    function getTimeline(uint fromId) public view returns (uint[] memory) {
-        uint count = 0;
-        uint cursor = fromId;
+    function getTimeline(uint256 fromId) public view returns (uint256[] memory) {
+        uint256 count = 0;
+        uint256 cursor = fromId;
 
         while (cursor != 0 && count < MAX_TIMELINE_DEPTH) {
             count++;
             cursor = nodes[cursor].previous;
         }
 
-        uint[] memory chain = new uint[](count);
+        uint256[] memory chain = new uint256[](count);
         cursor = fromId;
-        for (uint i = 0; i < count; i++) {
+        for (uint256 i = 0; i < count; i++) {
             chain[i] = cursor;
             cursor = nodes[cursor].previous;
         }
@@ -209,45 +194,45 @@ contract Universe is IUniverse, ReentrancyGuard, Pausable {
         return chain;
     }
 
-    function getLeaves() public view returns (uint[] memory) {
-        uint[] memory temp = new uint[](latestNodeId);
-        uint count = 0;
+    function getLeaves() public view returns (uint256[] memory) {
+        uint256[] memory temp = new uint256[](latestNodeId);
+        uint256 count = 0;
 
-        for (uint i = 1; i <= latestNodeId; i++) {
+        for (uint256 i = 1; i <= latestNodeId; i++) {
             if (nodes[i].id != 0 && nodes[i].next.length == 0) {
                 temp[count] = i;
                 count++;
             }
         }
 
-        uint[] memory leaves = new uint[](count);
-        for (uint j = 0; j < count; j++) {
+        uint256[] memory leaves = new uint256[](count);
+        for (uint256 j = 0; j < count; j++) {
             leaves[j] = temp[j];
         }
         return leaves;
     }
 
     /// @notice Paginated version of getLeaves() (L1 fix)
-    function getLeavesPage(uint256 offset, uint256 limit) public view returns (uint[] memory) {
+    function getLeavesPage(uint256 offset, uint256 limit) public view returns (uint256[] memory) {
         // First pass: count all leaves
         uint256 count = 0;
-        for (uint i = 1; i <= latestNodeId; i++) {
+        for (uint256 i = 1; i <= latestNodeId; i++) {
             if (nodes[i].id != 0 && nodes[i].next.length == 0) {
                 count++;
             }
         }
 
-        if (offset >= count) return new uint[](0);
+        if (offset >= count) return new uint256[](0);
 
         uint256 end = offset + limit;
         if (end > count) end = count;
         uint256 resultSize = end - offset;
 
-        uint[] memory result = new uint[](resultSize);
+        uint256[] memory result = new uint256[](resultSize);
         uint256 leafIndex = 0;
         uint256 resultIndex = 0;
 
-        for (uint i = 1; i <= latestNodeId && resultIndex < resultSize; i++) {
+        for (uint256 i = 1; i <= latestNodeId && resultIndex < resultSize; i++) {
             if (nodes[i].id != 0 && nodes[i].next.length == 0) {
                 if (leafIndex >= offset) {
                     result[resultIndex] = i;
@@ -260,14 +245,14 @@ contract Universe is IUniverse, ReentrancyGuard, Pausable {
         return result;
     }
 
-    function getMedia(uint id) public view returns (bytes32) {
+    function getMedia(uint256 id) public view returns (bytes32) {
         return nodes[id].contentHash;
     }
 
     /// @notice UNIVERSE-01: setMedia is now restricted to the original creator OR admin
     ///         on non-canon nodes only. Canon content is immutable on-chain — admin cannot
     ///         rewrite media that has been promoted to the canonical chain.
-    function setMedia(uint id, bytes32 _contentHash, string calldata _link) public {
+    function setMedia(uint256 id, bytes32 _contentHash, string calldata _link) public {
         require(nodes[id].id != 0, "Node does not exist");
         address originalCreator = nodes[id].creator;
         require(
@@ -284,11 +269,13 @@ contract Universe is IUniverse, ReentrancyGuard, Pausable {
     ///         canon integrity.
     /// @param nodeA First node ID
     /// @param nodeB Second node ID
-    function swapNodes(uint nodeA, uint nodeB) public onlyAdmin {
+    function swapNodes(uint256 nodeA, uint256 nodeB) public onlyAdmin {
         require(nodeA != nodeB, "Cannot swap a node with itself");
         require(nodes[nodeA].id != 0, "Node A does not exist");
         require(nodes[nodeB].id != 0, "Node B does not exist");
-        require(!nodes[nodeA].canon && !nodes[nodeB].canon, "UNIVERSE-01: canon nodes are immutable");
+        require(
+            !nodes[nodeA].canon && !nodes[nodeB].canon, "UNIVERSE-01: canon nodes are immutable"
+        );
 
         bytes32 tempContentHash = nodes[nodeA].contentHash;
         bytes32 tempPlotHash = nodes[nodeA].plotHash;
@@ -305,16 +292,12 @@ contract Universe is IUniverse, ReentrancyGuard, Pausable {
         emit NodesSwapped(nodeA, nodeB, msg.sender);
     }
 
-    function setNodeVisibilityOption(
-        NodeVisibilityOptions _option
-    ) public onlyAdmin {
+    function setNodeVisibilityOption(NodeVisibilityOptions _option) public onlyAdmin {
         nodeVisibilityOption = _option;
         emit NodeVisibilityOptionUpdated(_option);
     }
 
-    function setNodeCreationOption(
-        NodeCreationOptions _option
-    ) public onlyAdmin {
+    function setNodeCreationOption(NodeCreationOptions _option) public onlyAdmin {
         nodeCreationOption = _option;
         emit NodeCreationOptionUpdated(_option);
     }
@@ -323,25 +306,25 @@ contract Universe is IUniverse, ReentrancyGuard, Pausable {
         public
         view
         returns (
-            uint[] memory ids,
+            uint256[] memory ids,
             bytes32[] memory contentHashes,
             bytes32[] memory plotHashes,
-            uint[] memory previousIds,
-            uint[][] memory nextIds,
+            uint256[] memory previousIds,
+            uint256[][] memory nextIds,
             bool[] memory canonFlags
         )
     {
         require(latestNodeId <= 500, "Use getGraphPage for large graphs");
-        uint total = latestNodeId;
+        uint256 total = latestNodeId;
 
-        ids = new uint[](total);
+        ids = new uint256[](total);
         contentHashes = new bytes32[](total);
         plotHashes = new bytes32[](total);
-        previousIds = new uint[](total);
-        nextIds = new uint[][](total);
+        previousIds = new uint256[](total);
+        nextIds = new uint256[][](total);
         canonFlags = new bool[](total);
 
-        for (uint i = 1; i <= total; i++) {
+        for (uint256 i = 1; i <= total; i++) {
             VideoNode storage n = nodes[i];
 
             ids[i - 1] = n.id;
@@ -350,9 +333,9 @@ contract Universe is IUniverse, ReentrancyGuard, Pausable {
             previousIds[i - 1] = n.previous;
             canonFlags[i - 1] = n.canon;
 
-            uint len = n.next.length;
-            uint[] memory tmpNext = new uint[](len);
-            for (uint j = 0; j < len; j++) {
+            uint256 len = n.next.length;
+            uint256[] memory tmpNext = new uint256[](len);
+            for (uint256 j = 0; j < len; j++) {
                 tmpNext[j] = n.next[j];
             }
             nextIds[i - 1] = tmpNext;
@@ -364,33 +347,40 @@ contract Universe is IUniverse, ReentrancyGuard, Pausable {
     /// @notice Paginated version of getFullGraph. Returns nodes from startId to startId+count-1.
     /// @param startId First node ID to include (must be >= 1)
     /// @param count   Maximum number of nodes to return
-    function getGraphPage(uint startId, uint count)
+    function getGraphPage(uint256 startId, uint256 count)
         public
         view
         returns (
-            uint[] memory ids,
+            uint256[] memory ids,
             bytes32[] memory contentHashes,
             bytes32[] memory plotHashes,
-            uint[] memory previousIds,
-            uint[][] memory nextIds,
+            uint256[] memory previousIds,
+            uint256[][] memory nextIds,
             bool[] memory canonFlags
         )
     {
-        uint endId = startId + count - 1;
+        uint256 endId = startId + count - 1;
         if (endId > latestNodeId) endId = latestNodeId;
         if (startId > endId || startId == 0) {
-            return (new uint[](0), new bytes32[](0), new bytes32[](0), new uint[](0), new uint[][](0), new bool[](0));
+            return (
+                new uint256[](0),
+                new bytes32[](0),
+                new bytes32[](0),
+                new uint256[](0),
+                new uint256[][](0),
+                new bool[](0)
+            );
         }
-        uint total = endId - startId + 1;
+        uint256 total = endId - startId + 1;
 
-        ids = new uint[](total);
+        ids = new uint256[](total);
         contentHashes = new bytes32[](total);
         plotHashes = new bytes32[](total);
-        previousIds = new uint[](total);
-        nextIds = new uint[][](total);
+        previousIds = new uint256[](total);
+        nextIds = new uint256[][](total);
         canonFlags = new bool[](total);
 
-        for (uint i = 0; i < total; i++) {
+        for (uint256 i = 0; i < total; i++) {
             VideoNode storage n = nodes[startId + i];
             ids[i] = n.id;
             contentHashes[i] = n.contentHash;
@@ -398,9 +388,9 @@ contract Universe is IUniverse, ReentrancyGuard, Pausable {
             previousIds[i] = n.previous;
             canonFlags[i] = n.canon;
 
-            uint len = n.next.length;
-            uint[] memory tmpNext = new uint[](len);
-            for (uint j = 0; j < len; j++) {
+            uint256 len = n.next.length;
+            uint256[] memory tmpNext = new uint256[](len);
+            for (uint256 j = 0; j < len; j++) {
                 tmpNext[j] = n.next[j];
             }
             nextIds[i] = tmpNext;
@@ -415,12 +405,12 @@ contract Universe is IUniverse, ReentrancyGuard, Pausable {
     /// @dev The `canon` boolean on VideoNode is ONLY meaningful for `currentCanonId` — it marks the
     ///      current tip, not membership in the canon chain. Use getCanonChain() to get the full chain.
     ///      This is an O(1) operation — walk-and-update would be O(n) and gas-prohibitive for long chains.
-    function setCanon(uint id) public onlyAdmin {
+    function setCanon(uint256 id) public onlyAdmin {
         if (nodes[id].id == 0) {
             revert NodeDoesNotExist();
         }
 
-        uint previousCanon = currentCanonId;
+        uint256 previousCanon = currentCanonId;
 
         // O(1) — unset previous canon, set new one
         if (previousCanon != 0) {
@@ -431,7 +421,7 @@ contract Universe is IUniverse, ReentrancyGuard, Pausable {
         emit CanonChanged(id, previousCanon, msg.sender);
     }
 
-    function getCanonChain() public view returns (uint[] memory) {
+    function getCanonChain() public view returns (uint256[] memory) {
         if (currentCanonId == 0) {
             revert CanonNotSet();
         }
@@ -439,21 +429,23 @@ contract Universe is IUniverse, ReentrancyGuard, Pausable {
     }
 
     function getToken() public view returns (address) {
-      return associatedToken;
+        return associatedToken;
     }
 
-    function setToken(address token) external onlyManager{
+    function setToken(address token) external onlyManager {
         require(token != address(0), "Zero token address");
         associatedToken = token;
         emit TokenUpdated(token);
     }
+
     function setAdmin(address newAdmin) public onlyManager {
-      require(newAdmin != address(0), "Zero admin address");
-      universeAdmin = newAdmin;
-      emit AdminUpdated(newAdmin);
+        require(newAdmin != address(0), "Zero admin address");
+        universeAdmin = newAdmin;
+        emit AdminUpdated(newAdmin);
     }
-    function getAdmin() external view returns(address) {
-      return universeAdmin;
+
+    function getAdmin() external view returns (address) {
+        return universeAdmin;
     }
 
     // ---- Pausable (emergency stop) ----
