+++ apps/contracts/src/interfaces/IUniverse.sol
@@ -10,11 +10,13 @@ interface IUniverse {
     error CallerNotManager();
     error CallerNotAdmin(address caller);
 
-    event NodeCanonized(uint id, address canonizer); // deprecated — use CanonChanged
-    event CanonChanged(uint indexed newCanonId, uint indexed previousCanonId, address canonizer);
+    event NodeCanonized(uint256 id, address canonizer); // deprecated — use CanonChanged
+    event CanonChanged(
+        uint256 indexed newCanonId, uint256 indexed previousCanonId, address canonizer
+    );
     event NodeCreated(
-        uint indexed id,
-        uint indexed previous,
+        uint256 indexed id,
+        uint256 indexed previous,
         address indexed creator,
         bytes32 contentHash,
         bytes32 plotHash,
@@ -23,9 +25,14 @@ interface IUniverse {
     );
     event NodeVisibilityOptionUpdated(NodeVisibilityOptions option);
     event NodeCreationOptionUpdated(NodeCreationOptions option);
-    event MediaUpdated(uint indexed nodeId, address updater, bytes32 contentHash, string link);
-    event MediaUpdatedAttribution(uint indexed nodeId, address indexed updater, address indexed originalCreator, bytes32 contentHash);
-    event NodesSwapped(uint indexed nodeA, uint indexed nodeB, address swapper);
+    event MediaUpdated(uint256 indexed nodeId, address updater, bytes32 contentHash, string link);
+    event MediaUpdatedAttribution(
+        uint256 indexed nodeId,
+        address indexed updater,
+        address indexed originalCreator,
+        bytes32 contentHash
+    );
+    event NodesSwapped(uint256 indexed nodeA, uint256 indexed nodeB, address swapper);
     event WhitelistedUpdated(address indexed user, bool status);
     event VaultWhitelistUpdated(address indexed user, bool status);
     event TokenUpdated(address indexed token);
